Rename parsers and add it into alfredparser
